♻️[#60] Refactor: 파일 분리 완료
